Initial pass working
Simulates correctly. Has an example of a workable crafting
rotation. Uses brute-force DFS to explore the configuration space; I
did some experimentation for success-rate; it isn't even close to
being able to produce a solution from scratch, but being able to fill
in the last four or five actions in a rotation is a good start.